[Lab2] Remove workaround for qemu & use pyserial for qemu as well
Before this commit, we used a workaround in MiniUart.c to prevent
messing up the screen by replacing \r with \n for qemu env. This hurts
the functionality of data transfer. As a result, this commit removes
that workaround and uses pyserial to tackle that issue.

P.S. Adding timeout in pyserial is to prevent it hanging there for not
receiving the hello msg from rpi3.